Move questions into their own controller
- Nest ResearchSessions::QuestionsController and tell routes about it
- Move question step views to app/views/research_sessions/questions
- Move shared partials to app/views/shared
- Don't use implicit Wicked::Wizard variables like wizard_steps or
  step without explicit defined? checks
- Mix Barnardos FormHelper in at ApplicationController level
