+++ app/controllers/application_controller.rb
@@ -1,3 +1,5 @@
 class ApplicationController < ActionController::Base
+  helper Barnardos::ActionView::FormHelper
+
   protect_from_forgery with: :exception
 end
