+++ app/views/layouts/application.html.erb
@@ -5,7 +5,13 @@
   <meta charset="utf-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
-  <title><%= title(@research_session, step) %></title>
+  <title>
+    <% if defined?(step) %>
+      <%= title(@research_session, step) %>
+    <% else %>
+      <%= title(@research_session) %>
+  <% end %>
+  </title>
   <%= csrf_meta_tags %>
   <%= stylesheet_link_tag 'application', media: 'all' %>
   <%= stylesheet_link_tag 'print', media: 'print' %>
@@ -26,7 +32,7 @@
 </header>
 
 <div class="global-header-bar">
-  <%= render 'share' if @research_session.name %>
+  <%= render 'shared/share' if @research_session&.name %>
 </div>
 
 <main id="main" role="main">
