+++ app/views/research_sessions/_progress.html.erb
@@ -10,13 +10,13 @@
       give the user a link to that step based on if
       they have already visited it or not.
     %>
-    <% wizard_steps.each do |this_step| %>
+    <% ResearchSession::Steps.instance.step_keys.each do |this_step| %>
       <li
         class="progress__stage
           <%= 'is-complete' if @research_session.reached_step?(this_step) %>
-          <%= 'is-active' if this_step == step %>">
+          <%= 'is-active' if defined?(step) && this_step == step %>">
         <span class="progress__stage-title">
-          <% if this_step != step && @research_session.reached_step?(this_step) %>
+          <% if preview || (defined?(step) && this_step != step && @research_session.reached_step?(this_step)) %>
             <%= link_to this_step.to_s.humanize, research_session_question_path(@research_session.slug, this_step) %>
           <% else %>
             <%= this_step.to_s.humanize %>
@@ -30,10 +30,10 @@
       a link to the preview section
     %>
     <li class="progress__stage
-          <%= 'is-complete' if @research_session.reached_step?(wizard_steps.last) %>
+          <%= 'is-complete' if @research_session.reached_step?(ResearchSession::Steps.instance.step_keys.last) %>
           <%= 'is-active' if preview %>">
       <span class="progress__stage-title">
-        <% if !preview && @research_session.reached_step?(wizard_steps.last) %>
+        <% if !preview && @research_session.reached_step?(ResearchSession::Steps.instance.step_keys.last) %>
           <%= link_to t('application.preview_step_name'), research_session_preview_path(@research_session.slug) %>
         <% else %>
           <%= t('application.preview_step_name') %>
