+++ app/views/research_sessions/questions/_button_bar.erb
