+++ app/views/research_sessions/questions/topic.html.erb
@@ -1,6 +1,6 @@
 <%= render "research_sessions/progress" %>
 <div class="main-content">
-  <%= render 'error_summary' %>
+  <%= render 'shared/error_summary' %>
   <h1 class="subtitle-large">Topic</h1>
   <%= barnardos_form_with model: @research_session, url: wizard_path do |form| %>
     <%= form.labelled_text_field(
